i think backend is done

+++ nestjs/src/chat/chat.service.ts
@@ -7,6 +7,7 @@ import { Socket } from 'socket.io';
 import e from 'express';
 import { subscribe } from 'diagnostics_channel';
 import { ChatGateway } from './chat.gateway';
+import * as argon from 'argon2'
 
 @Injectable()
 export class ChatService {
@@ -32,18 +33,29 @@ export class ChatService {
 		return message;
 	}
 
-	async joinRoom(payload: any, client: Socket) {
+
+	//payload depan user_id, room_id
+	async joinRoom(payload: any) {
 		// console.log(payload);
 		const room = await this.prisma.chatRoom.findUnique({
 			where: {
 				id: payload.rid,
 			},
-			include: {
-				participants: true,
-			},
 		})
 		if (!room) {
 			throw new NotFoundException('Chat room not found');
+		}
+		if (room.visibility === 'PROTECTED') {
+			if (!payload.password) {
+				throw new NotFoundException('Password is required in protected room');
+			}
+			const isMatch = await argon.verify(room.password, payload.password);
+			if (!isMatch) {
+				throw new UnauthorizedException('Wrong password');
+			}
+		}
+		if (room.visibility === 'PRIVATE') {
+			throw new UnauthorizedException('Private room');
 		}
 		const user = await this.prisma.user.findUnique({
 			where: {
@@ -58,27 +70,13 @@ export class ChatService {
 			data: {
 				uid: payload.uid,
 				rid: payload.rid,
+				role: 'USER',
 			},
 		});
-		// const participant = this.prisma.participant.findUnique({
-		// 	where: {
-		// 		uid_rid: {
-		// 		uid: payload.id,
-		// 		rid: payload.rid,
-		// 		},
-		// 	},
-		// });
-		// if (!participant) {
-		// 	throw new NotFoundException('User not found in chat room');
-		// }
-		// client.join(payload.rid);
-		// client.to(payload.rid).emit('message', { msg : 'joined', room: payload.rid });
-
-		// client.to(payload.rid).emit('joined', payload);
 		return 'Joined room';
 	}
 
-	async kickUser(payload: any, client: Socket) {
+	async kickUser(payload: any) {
 		const room = await this.prisma.chatRoom.findUnique({
 		where: {
 			id: payload.rid,
@@ -101,6 +99,9 @@ export class ChatService {
 		if (!participant) {
 			throw new NotFoundException('User not found in chat room');
 		}
+		if (participant.role === 'OWNER') {
+			throw new UnauthorizedException('Cannot kick owner');
+		}
 		// console.log(payload);
 		await this.prisma.participant.delete({
 			where: {
@@ -115,16 +116,17 @@ export class ChatService {
 		}
 
 		// this.map.get(payload.id).leave(payload.rid);
-		client.to(payload.rid).emit('message', { msg : 'kicked', room: payload.rid });
+		// client.to(payload.rid).emit('message', { msg : 'kicked', room: payload.rid });
 		return 'Kicked user';
 	}
 
-
-	async createRoom(payload: any, client: Socket) {
-		console.log(payload);
+	// payload depan user_id, room_name, visibility, password if protected
+	async createRoom(payload: any) {
+		// console.log(payload);
 		const room = await this.prisma.chatRoom.create({
 			data: {
 				name: payload.name,
+				is_DM: false,
 			},
 		});
 		const participant = await this.prisma.participant.create({
@@ -134,16 +136,41 @@ export class ChatService {
 				role: 'OWNER',
 			},
 		});
+		if (payload.visibility === 'PROTECTED') {
+			if (!payload.password) {
+				throw new NotFoundException('Password is required in protected room');
+			}
+			const hash = await argon.hash(payload.password);
+			await this.prisma.chatRoom.update({
+				where: {
+					id: room.id,
+				},
+				data: {
+					visibility: 'PROTECTED',
+					password: hash,
+				},
+			});
+		} else if (payload.visibility === 'PRIVATE') {	
+			await this.prisma.chatRoom.update({
+				where: {
+					id: room.id,
+				},
+				data: {
+					visibility: 'PRIVATE',
+				},
+			});
+		}
 		return room;
 	}
 
-	async banUser(payload: any, client: Socket) {
+	//payload depan user_id, room_id
+	async banUser(payload: any) {
 		// console.log(payload.rid);
 		// client.leave(payload.rid);
 		const participant = await this.prisma.participant.findUnique({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -151,10 +178,13 @@ export class ChatService {
 		if (!participant) {
 			throw new NotFoundException('User not found in chat room');
 		}
+		if (participant.role === 'OWNER') {
+			throw new UnauthorizedException('Cannot ban owner');
+		}
 		await this.prisma.participant.update({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -162,17 +192,18 @@ export class ChatService {
 				isBanned: true,
 			},
 		});
-		this.map.get(payload.id).leave(payload.rid);
-		client.to(payload.rid).emit('message', { msg : 'banned', room: payload.rid });
+		// console.log('parti', participant);
+		// this.map.get(payload.uid).leave(payload.rid);
+		// client.to(payload.rid).emit('message', { msg : 'banned', room: payload.rid });
 		return 'Banned user';
 	}
 
-	async unbanUser(payload: any, client: Socket) {
+	async unbanUser(payload: any) {
 		// console.log(payload.rid);
 		const participant = await this.prisma.participant.findUnique({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -183,7 +214,7 @@ export class ChatService {
 		await this.prisma.participant.update({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -191,8 +222,8 @@ export class ChatService {
 				isBanned: false,
 			},
 		});
-		this.map.get(payload.id).join(payload.rid);
-		client.to(payload.rid).emit('message', { msg : 'unbanned', room: payload.rid });
+		// this.map.get(payload.id).join(payload.rid);
+		// client.to(payload.rid).emit('message', { msg : 'unbanned', room: payload.rid });
 		return 'Unbanned user';
 	}
 
@@ -213,7 +244,7 @@ export class ChatService {
 		const participant = await this.prisma.participant.findUnique({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -224,17 +255,17 @@ export class ChatService {
 		await this.prisma.participant.delete({
 			where: {
 				uid_rid: {
-					uid: payload.id,
+					uid: payload.uid,
 					rid: payload.rid,
 				},
 			},
 		});
-		client.leave(payload.rid);
 		client.to(payload.rid).emit('message', { msg : 'left', room: payload.rid });
+		client.leave(payload.rid);
 		return 'Left room';
 	}
 
-	async deleteRoom(payload: any, client: Socket) {
+	async deleteRoom(payload: any) {
 		// console.log(payload.rid);
 		const room = await this.prisma.chatRoom.findUnique({
 			where: {
@@ -255,12 +286,13 @@ export class ChatService {
 		return 'Deleted room';
 	}
 
-	async muteUser(payload: any, client: Socket) {
+	async muteUser(payload: any) {
 		// console.log(payload.rid);
+		
 		const participant = await this.prisma.participant.findUnique({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -268,10 +300,13 @@ export class ChatService {
 		if (!participant) {
 			throw new NotFoundException('User not found in chat room');
 		}
+		if (participant.role === 'OWNER') {
+			throw new UnauthorizedException('Cannot mute owner');
+		}
 		await this.prisma.participant.update({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -279,16 +314,16 @@ export class ChatService {
 				isMuted: true,
 			},
 		});
-		client.to(payload.rid).emit('message', { msg : 'muted', room: payload.rid });
+		// client.to(payload.rid).emit('message', { msg : 'muted', room: payload.rid });
 		return 'Muted user';
 	}
 
-	async unmuteUser(payload: any, client: Socket) {
+	async unmuteUser(payload: any) {
 		// console.log(payload.rid);
 		const participant = await this.prisma.participant.findUnique({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -299,7 +334,7 @@ export class ChatService {
 		await this.prisma.participant.update({
 			where: {
 				uid_rid: {
-				uid: payload.id,
+				uid: payload.uid,
 				rid: payload.rid,
 				},
 			},
@@ -307,20 +342,121 @@ export class ChatService {
 				isMuted: false,
 			},
 		});
-		client.to(payload.rid).emit('message', { msg : 'unmuted', room: payload.rid });
+		// client.to(payload.rid).emit('message', { msg : 'unmuted', room: payload.rid });
 		return 'Unmuted user';
 	}
 
+	async getAllRoom() {
+		const room = await this.prisma.chatRoom.findMany({
+			include: {
+				participants: true,
+			},
+		});
+		return room;
+	}
 
-  // findOne(id: number) {
-  //   return `This action returns a #${id} chat`;
-  // }
+	async getMessages(payload: any) {
+		const message = await this.prisma.message.findMany({
+			where: {
+				rid: payload.rid,
+			},
+			include: {
+				user: true,
+			},
+		});
+		// console.log({ message });
+		return message;
+	}
 
-  // update(id: number, updateChatDto: UpdateChatDto) {
-  //   return `This action updates a #${id} chat`;
-  // }
+	async changeVisibility(payload: any) {
+		const room = await this.prisma.chatRoom.findUnique({
+			where: {
+				id: payload.rid,
+			},
+		});
+		if (!room) {
+			throw new NotFoundException('Chat room not found');
+		}
+		if (room.visibility === 'PROTECTED') {
+			if (!payload.password) {
+				throw new NotFoundException('Password is required in protected room');
+			}
+			const isMatch = await argon.verify(room.password, payload.password);
+			if (!isMatch) {
+				throw new UnauthorizedException('Wrong password');
+			}
+		}
+		if (payload.visibility === 'PROTECTED') {
+			if (!payload.password) {
+				throw new NotFoundException('Password is required in protected room');
+			}
+			const hash = await argon.hash(payload.password);
+			await this.prisma.chatRoom.update({
+				where: {
+					id: payload.rid,
+				},
+				data: {
+					visibility: 'PROTECTED',
+					password: hash,
+				},
+			});
+		} else {
+			await this.prisma.chatRoom.update({
+				where: {
+					id: payload.rid,
+				},
+				data: {
+					visibility: payload.visibility,
+				},
+			});
+		}
+		return 'Changed visibility';
+	}
 
-  // remove(id: number) {
-  //   return `This action removes a #${id} chat`;
-  // }
+	async changeRoomName(payload: any) {
+		const room = await this.prisma.chatRoom.findUnique({
+			where: {
+				id: payload.rid,
+			},
+		});
+		if (!room) {
+			throw new NotFoundException('Chat room not found');
+		}
+		await this.prisma.chatRoom.update({
+			where: {
+				id: payload.rid,
+			},
+			data: {
+				name: payload.name,
+			},
+		});
+		return 'Changed room name';
+	}
+
+	async changePassword(payload: any) {
+		const room = await this.prisma.chatRoom.findUnique({
+			where: {
+				id: payload.rid,
+			},
+		});
+		if (!room) {
+			throw new NotFoundException('Chat room not found');
+		}
+		if (room.visibility !== 'PROTECTED') {
+			throw new NotFoundException('Room is not protected');
+		}
+		if (!payload.password) {
+			throw new NotFoundException('Password is required in protected room');
+		}
+		const hash = await argon.hash(payload.password);
+		await this.prisma.chatRoom.update({
+			where: {
+				id: payload.rid,
+			},
+			data: {
+				password: hash,
+			},
+		});
+		return 'Changed password';
+	}
 }
